Extract BrowserAddressBarCoordinator and split AppDelegate helpers
Pulls browser address-bar + omnibar-selection-repeat state out of
AppDelegate into a new @MainActor @Observable coordinator so the focused
panel ID is owned by one type and every NotificationCenter observer
callback enters a typed MainActor.assumeIsolated boundary before
touching it. Kills the cluster of Swift 6 "cannot be mutated from a
Sendable closure" and "main actor-isolated call from nonisolated
context" warnings that were stacking on AppDelegate.

Also extracts two self-contained helpers that were inlined in
AppDelegate.swift into their own files: VSCodeServeWebController plus
the code-tunnel serve-web launch helpers (VSCodeServeWeb.swift), and
CmuxCLIPathInstaller (CmuxCLIPathInstaller.swift). No behavior change
for either.

Net effect on AppDelegate.swift: 14,778 -> 13,785 lines (-993).
